Added Solar Export configuration option
Allows the enabling / disabling of the ability to sell excess solar. Users may want to control this option, as when enabled, the system is known to draw 40w at all times. The scenario here is to turn it off when solar is not expected to produce, and turn it back on when solar is producing.