Device based discovery & split ESP add-on (#493)
HomeAssistant MQTT integration (mqtt-entity): Use the "newer" device based discovery. 
Entity based discovery will be migrated. A drawback of device-based discovery is that you need to explicitly remove an entity from HA when you stop exposing that sensor from the sunsynk add-on (it becomes unavailable). This is fairly easy to do, and you can even select multiple unavailable entities to delete at the same time. Likely a better option than simply removing an entity.

Split the ESP sensor to an add-on